jobs/fetch-universe-metadata/requirements.txt
requests==2.32.3
azure-storage-blob==12.16.0
azure-identity==1.16.0